MultiFidelityGP Model Input Standardization (#482)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/482

We want to extend model input standardization to include multi fidelity models.

Previously, we finalized input standardization for `SingleTaskGP` and `FixedNoiseGP`. These two models only required `TrainingData` as inputs to `construct_inputs()`. However, the `...MultiFidelityGP` classes require `fidelity_features` on top of `TrainingData` as inputs.

To accomplish this, we:
- modify `construct_inputs` in `models/model.py` to include `fidelity_features` as a kwarg
- pass `fidelity_features` into `construct_inputs` for `SingleTaskGP` and `FixedNoiseGP`, but do not use the kwarg
- implement `construct_inputs` for the `...MultiFidelityGP` classes

Reviewed By: lena-kashtelyan

Differential Revision: D22467526

fbshipit-source-id: 35303aef9bdcc9f1c1139ce43354b1ecfc9ef16e