Allow conda build to use the setuptools_scm generated version number
Unfortunately `conda build`'s `load_setup_py_data` does not properly resolve the version number from `setuptools_scm`. The workaround here is to export this version number to an env variable that we can then substitute in to the jinja template of the `meta.yaml` file. This PR adds a simple shell script to make that easier, now the conda build process just amounts to calling `/build.sh` from the `.conda/` directory and the version number gets automatically added.

This also required some changes to `setup.py` to make sure the relative paths are properly resolved.