add seperate init_batch_limit option for generating initial_conditions
Summary: If batch_limit is too high, this can cause memory issues. However, batch_limit is used in optimize_acqf to limit the number of random restarts that we jointly optimize with LBFGS, which we often want to be low to reduce the dimension of the optimization problem. Using a super low batch_limit can make generating initial conditions unnecessarily slow.

Reviewed By: danielrjiang

Differential Revision: D24311761

fbshipit-source-id: 0a5134fab53878cda8fb497a7c662bf23ff76714