add seperate init_batch_limit option for generating initial_conditions (#586)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/586

If batch_limit is too high, this can cause memory issues. However, batch_limit is used in optimize_acqf to limit the number of random restarts that we jointly optimize with LBFGS, which we often want to be low to reduce the dimension of the optimization problem. Using a super low batch_limit can make generating initial conditions unnecessarily slow.

Reviewed By: danielrjiang

Differential Revision: D24311761

fbshipit-source-id: e1dcc0d131b1af37129799d2157b7709c50f7a0c