Add SMOKE_TEST parameters to tutorials (#692)
Summary:
Add a SMOKE_TEST variable to all tutorial notebooks that take a while to run, based on which dummy settings that significantly reduce the optimization quality / number of iterations are selected so that we can run these in CI in a reasonable time. SMOKE_TEST is read from `os.environ["SMOKE_TEST"], which set in the GitHub Actions workflow.

Pull Request resolved: https://github.com/pytorch/botorch/pull/692

Reviewed By: sdaulton

Differential Revision: D26485797

Pulled By: Balandat

fbshipit-source-id: 814034b52ab9904d1b2fb0b549675cc0cbe07394