Add "Test against stable" workflow.
This workflow tests against current releases of gpytorch/pytorch. Since these will generally fail it is not set up to execute automatically, only on manual dispatch.

This workflow is primarily intended to be used prior to releasing a new version to check compatibility with current releases (and without executing the full deployment workflow that pushes packages and the website).