Add "Test against stable" workflow. (#706)
Summary:
This workflow tests against current releases of gpytorch/pytorch. Since these will generally fail it is not set up to execute automatically, only on manual dispatch.

This workflow is primarily intended to be used prior to releasing a new version to check compatibility with current releases (and without executing the full deployment workflow that pushes packages and the website).

Pull Request resolved: https://github.com/pytorch/botorch/pull/706

Reviewed By: dme65

Differential Revision: D26547115

Pulled By: Balandat

fbshipit-source-id: 01816241b847d9da551cf20a8155c09464229927