Introduce base_sample_shape property to Posterior objects (#718)
Summary:
This allows base sample shapes that are different from the posterior's `event_shape`. This is relevant e.g if for GPyTorch MVN posteriors the covariance is singular, or if a low-rank approximation of the root decomposition is used.
See the corresponding GPyTorch PR here: https://github.com/cornellius-gp/gpytorch/pull/1502

Also does some cleanup and moves the `HigherOrderGPPosterior` into the `posteriors` module.

Pull Request resolved: https://github.com/pytorch/botorch/pull/718

Reviewed By: sdaulton

Differential Revision: D26608568

Pulled By: Balandat

fbshipit-source-id: 8b6cf11a99f1d9459fa4971ab318cf54e741f8bc