utility for converting batched mo models into batched so models (#759)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/759

This is helpful for several entropy acquisition functions in the works. This is also useful because it allows you to compute a batched independent MVN (with a covariance matrix of shape m x n x n) rather than the full MTMVN (with a covariance of shape mn x mn), which is far more memory efficient.

Reviewed By: Balandat

Differential Revision: D25816540

fbshipit-source-id: 842efce4e1ad74dab03a9dc59d5df990db6c92ff