utility for converting batched mo models into batched so models (#759)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/759

This is helpful for several entropy acquisition functions in the works. This is also useful because it allows you to compute a batched independent MVN (with a covariance matrix of shape m x n x n) rather than the full MTMVN (with a covariance of shape mn x mn), which is far more memory efficient.

Differential Revision: D25816540

fbshipit-source-id: c71f1438f17a3b8b5aee362ffeed8c9e3ba394a0