update default partitioning selection heuristic (#846)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/846

We now have a very fast non-dominated partitioning that is much more scalable with the number of objectives.

Using 100 data points on the PF (a very large PF):
For 3 objectives, FastNondominatedPartitioning (exact) is faster. For 4 objectives, FastNondominatedPartitioning (exact) is competitive with the approximate heuristic. For 5 objectives, FastNondominatedPartitioning is a good deal slower (still not that bad).

Differential Revision: D29353926

fbshipit-source-id: d98dc1928b032a8506a95e130f5cad6f2926228d