Fix qMFKG evaluate bug (#858)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/858

In `qMFKG.evalute` the use of `if hasattr(self, "cost_aware_utility")` leads to a bug when `cost_aware_utility=None` (ends up calling `None` as a function). This replaces that with `if getattr(self, "cost_aware_utility", None) is not None`, which handles both not having a `cost_aware_utility` attribute (qKG) and `cost_aware_utility=None` (qMFKG).

Reviewed By: Balandat

Differential Revision: D29565578

fbshipit-source-id: e3bed21a9796879dace1f95c910621d52518fa84