Support negative weights for minimization objectives in ParEGO (#884)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/884

In the previous implementation, all objectives are normalized to [0,1]. Maximization objectives are assigned random positive weights and minimization objectives are assigned random negative weights (in Ax) such that sum(w\*y) makes sense. However, this will make w\*y negative for the minimization objectives, which makes the min(w\*y) term in augmented Chebyshev scalarization always points to the scalarized minimization objectives and neglects the scalarized maximization objectives. This will cause issue in optimization, as we want the min(w\*y) term to focus on the scalarized objective that has small weight and hence neglected by the sum(w\*y) term. The fix is to shift the normalized minimization objectives from [0,1] to [-1,0]. The direction of the maximization is still preserved since it is a shift, and the weights are still going to be negative for minimization objectives. But now all w\*y's are in the positive range and min(w\*y) makes sense again.

Reviewed By: sdaulton

Differential Revision: D29945933

fbshipit-source-id: cd492db7679afff018bc7ee7e45f97e6d7ae0942