Introduce ModelList and PosteriorList (#829)
Summary:
This introudces a `ModelList` container for using heterogeneous models. This is very similar to `ModelListGP`, but more generic in the sense that it works with general objects that implement the `Model` API, not just GPs.

The particular use case I have in mind for this is to have a multi-output model where one or more outputs are deterministic functions of the input. Using `ModelList`, we can just throw some `DeterministicModel` objects into a `ModelList`, together with GP models. This is useful e.g. for doing Multi-Objective optimization when some objectives are deterministic functions of the input, which is the case in some Chemistry problems or in the context of optimizing sparsity-promoting policies. In that case, we should be able to simply stick the `ModelList` into existing MOO acquisition functions and everything should (famous last words) work.

Pull Request resolved: https://github.com/pytorch/botorch/pull/829

Reviewed By: bletham

Differential Revision: D29187914

Pulled By: Balandat

fbshipit-source-id: b579868ff47e993f74e20eb42dd1966c61987a34