Short-cut t_batch_mode_tarnsform decorator on non-tensor inputs
This essentially makes `t_batch_mode_tarnsform` a nullop in case the argument `X` passed to the acquisition funciton is not a `torch.Tensor` object. This allows using acquisition funcitons that use models with non non-standard input types such as strings, which is the case in some applicaitons.

Currently this just touches the decorator; in the future we should consider changing the types and signatures of the acqusiition functions and models throughout to natively support this more generally.