Short-cut t_batch_mode_transform decorator on non-tensor inputs (#991)
Summary:
This essentially makes `t_batch_mode_tarnsform` a nullop in case the argument `X` passed to the acquisition function is not a `torch.Tensor` object. This allows using acquisition functions that use models with non non-standard input types such as strings, which is the case in some applications.

Currently this just touches the decorator; in the future we should consider changing the types and signatures of the acquisition functions and models throughout to natively support this more generally.

cc wjmaddox

Pull Request resolved: https://github.com/pytorch/botorch/pull/991

Reviewed By: dme65

Differential Revision: D32903859

Pulled By: Balandat

fbshipit-source-id: c3abc8b40db307358807fe60014e2c0e1fe49c58