Fix find_interior_point for negative variables & allow unbounded problems (#1045)
Summary:
`scipy.optimize.linprog` has the default `bounds=(0, None)`, meaning that it by default assumes all variables to be positive. This will break things if the feasible set does not lie in the positive orthant. This PR fixes that.

It also introduces a fallback method for unbounded problems, in which case we now find a feasible solution with minimal L1 norm.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1045

Reviewed By: dme65

Differential Revision: D33270302

Pulled By: Balandat

fbshipit-source-id: 3e1cc4a9246bab1bbf856a23515b1fca76282abb