Improve find_interior_point (#1049)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1049

Modifies the LP for finding an interior point to better deal with unbounded feasible sets by constraining the slack variable.

This change is necessary since if in the existing version (which essentially searches for the point “furthest away from the boundary”) the feasible set is unbounded, then LP will find a solution on the boundary (the one with minimal inf norm) which is not an interior point. This PR gets rid of that logic and caps the “furthest away to” some (arbitrary) max so that we don’t run into issues with unboundedness in that case. This also simplifies things relative to the previous fallback. We could of course also just always solve the bounded version, but that might cause some weird outcomes, so I am doing this as a fallback only.

Reviewed By: dme65

Differential Revision: D33325860

fbshipit-source-id: 40000aff08e67787f414d6331895ad2737d65578