make non-interleaved base samples use intuitive shape (#1057)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1057

This was pretty subtle but pretty important. Previously, base_samples with shape: `sample_shape x n x m` would not actually be mapped to outcomes in the way that one would expect. Appending the base samples for a `n+1`th point would affect the posterior samples for the some of the `n` points due to not accounting for not interleaving. This update accounts for not interleaving, such that appending new base samples does not change the posterior samples for the previous points.

Reviewed By: dme65

Differential Revision: D33532706

fbshipit-source-id: 5548ac9c7ecd0149b95eb43ff5f4529fbe186d1d