Add support for non-linear input constraints (#1067)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1067

This makes it possible to add a list of non-linear input constraints that need to be satisfied when optimizing the acquisition function. These are expected to be passed down as a list of callables where each callable takes a torch tensor as an input. Internally, we convert these into the format expected by SLSQP and use autograd to compute the gradient of the constraint.

Reviewed By: Balandat

Differential Revision: D33551393

fbshipit-source-id: 742c4ff537cdece8b4932a53250236b698cb1cdd