low-rank cholesky updates for NEI (#1056)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1056

This uses low-rank cholesky updates in NEI. Using SAA this allows us to cache the objectives values for the in-sample points and only compute the objectives for the new test points. This is much faster when there are lots of baseline points.

This refactors components that are shared with NEHVI into a shared abstract class.

Reviewed By: Balandat

Differential Revision: D32668278

fbshipit-source-id: 79deb6606168ed3f5f432a232600d08e840c9bfb