Adding MOMF acquisition function (#1153)
Summary:
<!--
Thank you for sending the PR! We appreciate you spending the time to make BoTorch better.

Help us understand your motivation by explaining why you decided to make this change.

You can learn more about contributing to BoTorch here: https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md
-->

## Motivation

The MOMF acquisition function is targeted at simultaneously optimizing a Multi-objective Multi-Fidelity problem. It does this by making use of a 'Trust Objective' which is optimized along with the other multi-objectives. This allows the MOMF acquisition function to balance the trade-off of going for the cheaper objective evaluations vs yielding higher values for Trust objective by going to higher fidelity evaluations. This PR includes an implementation of the MOMF acquisition function, new Multi-objective Multi-Fidelity test functions and a tutorial jupyter notebook that walks through the procedure of using MOMF to optimize a Modified Branin-Currin problem. For more details on the working of the acquisition function see
[Irshad, F., Karsch, S. and Döpp, A., 2021. Expected hypervolume improvement for simultaneous multi-objective and multi-fidelity optimization. arXiv preprint arXiv:2112.13901].

### Have you read the [Contributing Guidelines on pull requests](https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md#pull-requests)?

Yes I have and tried to adhere to all the guidelines mentioned.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1153

Test Plan:
For the MOMF acquisition function based on qEHVI, I made use of some of the test cases of the qEHVI and rewrote some parts of it. The changes include new values for acquisition function values since the MOMF internally divides by the cost of evaluating points and so the expected values from the test cases were changed to account for this. I wrote a new test case as well where I control the fidelity cost and check if the MOMF returns the true accurate value within a certain tolerance. The MOMF is also checked for returning the right tensor shapes in case of q=1,2,3.
The test functions also undergo a similar check where they are checked for correct initialization and returning the optimized values when evaluated at optimizers.

## Related PRs

This does adds functionality and so I have created a jupyter tutorial notebook to outline this functionality. It is a part of this PR.
(If this PR adds or changes functionality, please take some time to update the docs at https://github.com/pytorch/botorch, and link to your PR here.)

Reviewed By: sdaulton

Differential Revision: D36398219

Pulled By: Balandat

fbshipit-source-id: 02b9711f650b017c46a4108451e4e4786b537d87