option to generate requirements.txt output
Summary:
This provides a simple `--requirements` flag to pyfmt, to output
a simple requirements file with the version of black/usort in use.
It also adds `fbsource/tools/lint/pyfmt/reqs/requirements-fmt.txt`
generated from the output of `pyfmt --requirements`.

This will allow open source projects with source of truth in fbsource
to export this directory via shipit, so that they automatically get
pinned to the active version of black/usort used internally.

The mechanism for ensuring this file stays in sync with reality
will be devised in a later diff.

sleepykirby

Reviewed By: zertosh

Differential Revision: D36431427

fbshipit-source-id: 7954f33947a72b98ea27500384d6ada5c83d36a7