PE and PLBO support in Ax
Summary:
Support PE/PLBO in Ax by revamping Shaun's previous implementation.

Given that now the preference data is stored as an Ax experiment, we no longer need to pass in the preference data around, but the `pref_model` instead. `pref_model` will either be passed directly into the EUBO acqf (as in the PE exmample) or will serve as the objective (as in the PLBO example), extracting it from a `PairwiseModelBridge` and passing it through `acquisition_options` seems to be a nice solution. Comments and feedbacks are welcomed.

This should be the last missing piece of supporting PL/PE/PLBO in Ax. Though we should still need to work out a usable NB UI (probably something based on top of the BoTorch backend PL NB game), but with this diff, we should be able to perform close-loop PL/PE/PLBO in Ax programmatically.

Reviewed By: lena-kashtelyan

Differential Revision: D36060116

fbshipit-source-id: ca2376481fd356bd7b2ad0cb3826171cc5daf521