Make `task_feature` as required input in MultiTaskGP.construct_inputs (#1246)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1246

Motivation:
We see an error in CV when using ListSurrogate (a list of MTGP) in Models.BOTORCH_MODULAR. See N2054031 for the full error trace.

The error happens when loading state_dict. The reason is that a wrong task_feature is passed to MTGP (use default value = 0) and thus, the task_covar_module has a wrong shape.

The proposed the fix here is: Make task_feature as required input in the MultiTaskGP.construct_inputs

Reviewed By: Balandat

Differential Revision: D36925653

fbshipit-source-id: 8cfde439083bb47d141f6cfc872646ffce7838de