Remove deprecated usage of setuptools_scm in setup.py
This removes the deprecated use of use_setuptools_scm in setup.py so that we can use newer setuptools versions. Previously we parsed the USE_SCM_LOCAL_VERSION env var as part of setup.py, but this is not possible in the pyproject.toml. Instead we now simply use the SETUPTOOLS_SCM_PRETEND_VERSION env var to provide a specific version name without the github hash so things can be uploaded to test.pypi.org