Remove deprecated usage of setuptools_scm in setup.py (#1292)
Summary:
This removes the deprecated use of `use_setuptools_scm` in `setup.py` so that we can use newer setuptools versions.

Previously we parsed the `USE_SCM_LOCAL_VERSION` env var as part of `setup.py` in order to programmatically set the `local_scheme` setting, but this kind of dynamism is not possible in the pyproject.toml configuration. Instead, we now simply use the `SETUPTOOLS_SCM_PRETEND_VERSION` env var in order to provide a specific version name without the github hash (which we strip from the full name in a bash script).

What this means is that for untagged commits, `setuptools_scm` will always create a version string that includes the github hash and date (and overriding this will require to manually specifying the `SETUPTOOLS_SCM_PRETEND_VERSION` env var).

Pull Request resolved: https://github.com/pytorch/botorch/pull/1292

Reviewed By: saitcakmak

Differential Revision: D37720650

Pulled By: Balandat

fbshipit-source-id: fdc92670086e9ca586da0d6260e5eb916ff90605