Skip bound validation in optimize_acqf if inequality constraints are specified (#1297)
Summary:
In some cases we may allow not using box constraints (e.g. when optimizing Alebo). This removes a check that currently breaks Alebo optimization. A proper solution should actually validate the constraint set, this was started in https://github.com/pytorch/botorch/pull/1231 but there are some nontrivial issues with this, so this provides a quick fix in the meantime.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1297

Reviewed By: saitcakmak

Differential Revision: D37773082

Pulled By: Balandat

fbshipit-source-id: cc64868a6c0af61aa3c846d2ed51da02cc5f6a20