Support inf values in `bounds` arg of `optimize_acqf` (#1302)
Summary:
Previously, the code assumed that finite bounds were provided. This allows specifying `inf` values which are then internally identified and passed as `None` to optimizers.

Note that in that case `batch_initial_conditions` will need to be provided since the default initializers will raise a `NotImplementedError` if `bounds` contain `inf` values. This is not a common scenario and this change is mainly to fix a current bug with Alebo on the Ax side. Down the road we may want to consider initialization strategies for spaces without finite (explicit) bounds if needed.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1302

Reviewed By: saitcakmak

Differential Revision: D37821301

Pulled By: Balandat

fbshipit-source-id: b5ef3a9ddca783e3e12ec7d970dd7c854aee135d