Implement MARS as a risk measure (#1303)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1303

Implements MARS as a BoTorch risk measure.

MARS uses a modified version of `get_chebyshev_scalarizations` that relies on a complex heuristic for selecting the normalization bounds. This does not work with negative weights (which need to be supported for Ax reasons), so we implement the modified scalarization as a method of MARS instead.

Reviewed By: Balandat

Differential Revision: D36467527

fbshipit-source-id: 122e62c72ae7782304d462a2b64f2b20d3d9dec7