Add a `skip_expand` option to AppendFeatures
Summary: Adds a `skip_expand` option to `AppendFeatures` transform, which skips the `expand` operation in the `forward` pass, allowing it to be mixed with an `InputPerturbation` transform without multiplying the size of the output tensor. When used together, each perturbation and each appended feature will appear only once in the output tensor per q-batch.

Differential Revision: D38594609

fbshipit-source-id: 5faf725fe004bddb8756f3e5d91bdf371ead841d