Add a `skip_expand` option to AppendFeatures (#1344)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1344

Adds a `skip_expand` option to `AppendFeatures` transform, which skips the `expand` operation in the `forward` pass, allowing it to be mixed with an `InputPerturbation` transform without multiplying the size of the output tensor. When used together, each perturbation and each appended feature will appear only once in the output tensor per q-batch.

Reviewed By: danielrjiang

Differential Revision: D38594609

fbshipit-source-id: 8cf1a8cbfb1be6faf3ba7a524d784f6488f461f0