Allow passing Y_samples directly in MARS.set_baseline_Y
Summary: We currently sample from the model to generate `Y` to compute the `MVaR` of, and set the Pareto subset of MVaR as the `baseline_Y`. This adds an option to skip the sampling bit and provide the `Y` directly. This saves us compute in cases where we may have access to pre-computed samples and may not want to re-sample.

Differential Revision: D38950165

fbshipit-source-id: 4ae4ef89b1ab5885fe93f5c89351315dfe027ca1