Allow passing Y_samples directly in MARS.set_baseline_Y (#1364)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1364

We currently sample from the model to generate `Y` to compute the `MVaR` of, and set the Pareto subset of MVaR as the `baseline_Y`. This adds an option to skip the sampling bit and provide the `Y` directly. This saves us compute in cases where we may have access to pre-computed samples and may not want to re-sample.

Reviewed By: esantorella

Differential Revision: D38950165

fbshipit-source-id: 8eb24b086a03ad0035e4018f54975d057eaa1c95