Deprecating GPyTorch's `NanError`, `NotPSDError`, `psd_safe_cholesky` in favor of LinearOperators's versions (#1377)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1377

Due to the recent migration to LinearOperators, we get a number of deprecation warnings, suggesting to switch from gpytorch versions of certain functions and objects to their LinearOperator equivalents. This diff aims to move all occurences of `NanError`, `NotPSDError`, `psd_safe_cholesky` in BoTorch to the new versions, thereby minimizing the number of deprecation warnings.

Reviewed By: Balandat

Differential Revision: D39287750

fbshipit-source-id: f2f291465ade1889a9988b5072cefe346f5acf5e