Deprecate `weights` argument of risk measures in favor of a `preprocessing_function`. (#1400)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1400

Deprecates the `weights` argument of risk measures in favor of a `preprocessing_function`. This is superior in that it allows better modification of the samples before computing the risk measures.

This supports use cases such as filtering non-objective outcomes or applying feasibility weighting all within the risk measure itself. As a result, it helps avoid a number of if/else blocks when implementing robust optimization support in Ax.

Reviewed By: Balandat

Differential Revision: D39493308

fbshipit-source-id: bf065e93237a805226687fec3637279afdd93065