Add a warning when using float dtype (#1445)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1445

We commonly see issues with users running into numerical errors when using float inputs. This adds a dtype check in `_validate_tensor_args`, which is used in the constructor of a large number of models, warning the users if they're not using double, and pointing them to https://github.com/pytorch/botorch/discussions/1444. Feel free to edit & add more details / examples to the discussion!

Reviewed By: Balandat

Differential Revision: D40199087

fbshipit-source-id: eb9f320e6872b0fab9b547534c9c43901d09ea1f