Fix input transform bug when sequentially training a `BatchedMultiOutputModel` (#1454)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1454

This fixes a bug where, due to the input transform being in `train` mode while the model is in `eval` mode, the learnable bounds of input transforms, such as Normalize, would be wrongly updated during model training, effectively disabling the transforms.

This is a serious bug that was leading to using the model trained with normalized inputs on a model without any normalization. It has been around for a while, since https://github.com/pytorch/botorch/pull/1283 :'(.

Differential Revision: D40453200

fbshipit-source-id: 0d60ac5f23efe06e200713555dfa109b612d9290