Adding `AffineInputTransform` (#1461)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1461

This diff adds `AffineInputTransform`, which applies `(x-b)/a` to an input `x`. Since this generalizes both `Normalize` and `Standardize`, this diff also makes them derived classes of the new class, leading to code sharing between the three classes.

Background: For concrete strength prediction, the `log(t + 1`) transformation of the time dimension greatly improves predictive accuracy and uncertainty callibration. It further holds promise for the learning curve prediction problem. Instead of directly implementing the custom transformation, I wrote this generalized affine transformation, which can be composed with `log` to yield the desired transformation.

Reviewed By: saitcakmak, Balandat

Differential Revision: D40043491

fbshipit-source-id: d6a8c9441ce21d442c095203df448506c0e49a9b