Introducing `subset_transform` decorator (#1468)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1468

Most `InputTransforms` have an `indices` field that specifies a feature subset on which to apply the transform. This diff introduces the `subset_transform` decorator, which separates out the indexing logic, thereby simplifying the implementation of mutliple `transform` and `untransform` methods, as well as adding support for `indices` in `InputPerturbation`.

Reviewed By: saitcakmak

Differential Revision: D40620269

fbshipit-source-id: 05fe093569d12bb19f9597044317efe270ba9355