Fix tutorials after introduction of MetricFetchResult
Summary:
A few tutorials were not compatible with the new way we fetch data. In this diff we change custom metrics to return Ok(data) instead of data, and change `trial.fetch_data()` => `Metric._unwrap_trial_data_multi(results=trial.fetch_data())`.

In the future we may want to add some copy in the tutorials discussing Result, but we should land this for now to get things working again ASAP

Differential Revision: D40994060

fbshipit-source-id: 0246818e55659090fdd12cd1082bc8d807c93b27