Do not count hitting maxiter as optimization failure & update default maxiter (#1478)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1478

We currently retry optimization upon hitting maxiter, which typically leads to optimizing the same thing twice. This diff updates the warning handling to avoid this behavior.

The default maxiter is also updated to 2_000, which is a steep change from scipy defaults of 15_000 for L-BFGS-B and 100 for SLSQP. This is just a temporary setting to bridge the gap between the two until we come up with better stopping conditions for the optimizers.

Reviewed By: esantorella

Differential Revision: D41053812

fbshipit-source-id: a0f27fb72f3cab6343fb55b6041717706e2445ce