Fix RFF bug when using FixedNoiseGP models
Summary: `_model.likelihood.noise` is a `... x d`-dim tensor for fixed noise models with `d`-dimensional `train_X`. This leads to errors when constructing the RFF. This updates the `sigma_sq` to use the mean across the input dimensions, which was used prior to https://github.com/pytorch/botorch/pull/1336.

Differential Revision: D41568680

fbshipit-source-id: fb9b422582cc57fb35859d5d3b20c92c140ee38b