Fix RFF bug when using FixedNoiseGP models
Summary: `_model.likelihood.noise` is a `... x d`-dim tensor for fixed noise models with `d`-dimensional `train_X`. This leads to errors when constructing the RFF. This updates the `sigma_sq` to use the mean across the input dimensions, which was used prior to https://github.com/pytorch/botorch/pull/1336.

Differential Revision: https://internalfb.com/D41568680

fbshipit-source-id: 7f846bc7b33d964f6574107219d1dabfeec5106f