Fix RFF bug when using FixedNoiseGP models (#1528)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1528

`_model.likelihood.noise` is a `... x d`-dim tensor for fixed noise models with `d`-dimensional `train_X`. This leads to errors when constructing the RFF. This updates the `sigma_sq` to use the mean across the input dimensions, which was used prior to https://github.com/pytorch/botorch/pull/1336.

Reviewed By: Balandat

Differential Revision: D41568680

fbshipit-source-id: fef740ebd62e49d63397ebc4ea1d1dc883f798aa