Increasing code sharing of `AnalyticAcquisitionFunctions` (#1558)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1558

This commit increases shared code between all `AnalyticAcquisitionFunction` classes by introducing a `_posterior_moments` function that combines universally used logic. It further adds `ImprovementAcquisitionFunction`, which unifies the constructor and specializes `_posterior_moments` for all improvement-based acquisition functions.

`_posterior_moments` also moves all parameters and buffers to the appropriate device, which was so far not the case for `ScalarizedPosteriorMean`, which also gets a `maximize` keyword argument to unify the interface with the other functions.

Reviewed By: Balandat

Differential Revision: D41880200

fbshipit-source-id: 5a1eb7ef9d3787e2a510fd3d3fc1cb2838c3e926