Introducing `LogExpectedImprovement` (#1565)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1565

This commit introduces `LogExpectedImprovement`, which computes the logarithm of the Expected Improvement acquisition function. By taking special care of numerics, it avoids numerical issues that affect `ExpectedImprovement` when evaluated at points that with high certainty do not yield improvement. In contrast, `LogExpectedImprovement` computes both values and gradients in a numerically stable fashion for points that don't yield improvement down to at least **100 orders of magnitude** lower posterior standard deviation than vanilla EI in double precision. This allows optimizers to escape the "valley of death" - regions that the surrogate is certain are bad but for which EI does not yield gradients.

On the example test problem in N2901527, using `logEI` leads to better convergence compared to GPEI and similar performance as running homotopy on the target value without having to run homotopy iterations.
{F822088495}

Reviewed By: Balandat

Differential Revision: D41890063

fbshipit-source-id: 368278afd4b2a84f66664070a70e326d0d66bedf