Introduce ModelDict and MultiModelAcquisitionFunction
Summary:
Introduces a lightweight `ModelDict` container, which is simply a `ModuleDict[str, Model]`, and an abstract `MultiModelAcquisitionFunction` class that accepts a `ModelDict` rather than a `Model`.

The goal here is to help shape the MBM Surrogate refactor by having a concrete example of how the multiple surrogates would be consumed in BoTorch.

Differential Revision: D41564744

fbshipit-source-id: dd1fec3db4cf851b3886407f1becf79ad25735fc