Introduce ModelDict and MultiModelAcquisitionFunction
Summary:
Introduces a lightweight `ModelDict` container, which is simply a `ModuleDict[str, Model]`, and an abstract `MultiModelAcquisitionFunction` class that accepts a `ModelDict` rather than a `Model`.

The goal here is to help shape the MBM Surrogate refactor by having a concrete example of how the multiple surrogates would be consumed in BoTorch.

Reviewed By: lena-kashtelyan

Differential Revision: D41564744

fbshipit-source-id: 1ff672a3c713e7213a87536588852973360dee99