Introduce ModelDict and MultiModelAcquisitionFunction (#1584)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1584

Introduces a lightweight `ModelDict` container, which is simply a `ModuleDict[str, Model]`, and an abstract `MultiModelAcquisitionFunction` class that accepts a `ModelDict` rather than a `Model`.

The goal here is to help shape the MBM Surrogate refactor by having a concrete example of how the multiple surrogates would be consumed in BoTorch.

Reviewed By: lena-kashtelyan

Differential Revision: D41564744

fbshipit-source-id: d18632b94041d3529cc0b98560d7791e7d83d59d