Add Constraint Active Search tutorial to ignore list (#1588)
Summary:
## Motivation

The tutorials have been failing in the nightly cron for several days. [Binary search](https://github.com/pytorch/botorch/pull/1581) identifies the constraint active search tutorial as the culprit. I suspect this is an OOM. I plan to fix it after the holidays, but since that will be over a week (optimistically) let's ignore it for now.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1588

Test Plan:
Manually trigger cron on this branch and make sure it passes.

## Related PRs

https://github.com/pytorch/botorch/issues/1581  debugs this.

Reviewed By: saitcakmak

Differential Revision: D42224423

Pulled By: esantorella

fbshipit-source-id: 10e4d619ef5d4c3ade46fba7f844caf9993c35ed