Add a minimize_with_timeout wrapper for scipy.optimize.minimize (#1403)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1403

Unfortunately, scipy does not allow timing out the optimization based on wall time. This diff implements a lightweight wrapper around `scipy.optimize.minimize` to achieve this.

The new `minimize_with_timeout` method calls `scipy.optimize.minimize` with all arguments forwarded verbatim. The only difference is that if provided a `timeout_sec` kwarg, it automatically stops the optimization after the timeout is reached.

Internally, this is achieved by automatically constructing a callback method that is injected to the `scipy.optimize.minimize` call that keeps track of the runtime and is used to extract the value of the optimization variables at the current iteration.

Reviewed By: esantorella

Differential Revision: D39529835

fbshipit-source-id: 76941d8e94d4d5d2946b13a7946a5523646c1185