Fix tensor shapes in unit tests
Summary: As a result of switching from `self.AssertTrue(torch.allclose(...))` to `self.AssertAllClose` in unit tests, we will now also have checks that tensors compared are the same shape and not just numerically equal. Some of our current tests were failing; this fixes that by changing the shapes of the compared tensors.

Differential Revision: D42402387

fbshipit-source-id: 032b17273c8aa4994f7e1ccff7b31040e51082f8