Fix tensor shapes in unit tests
Summary: As a result of switching from `self.AssertTrue(torch.allclose(...))` to `self.AssertAllClose` in unit tests, we will now also have checks that tensors compared are the same shape and not just numerically equal. Some of our current tests were failing; this fixes that by changing the shapes of the compared tensors.

Differential Revision: D42402387

fbshipit-source-id: c19aac04c09e0602ce953f045dd9e913295bb958