Fix tensor shapes in unit tests (#1620)
Summary:
Pull Request resolved: https://github.com/pytorch/botorch/pull/1620

As a result of switching from `self.AssertTrue(torch.allclose(...))` to `self.AssertAllClose` in unit tests, we will now also have checks that tensors compared are the same shape and not just numerically equal. Some of our current tests were failing; this fixes that by changing the shapes of the compared tensors.

Reviewed By: Balandat

Differential Revision: D42402387

fbshipit-source-id: 2a375b21d1c309142cfd40a231a5f5bf4861be5f