Speed up the SAASBO tutorial (#1740)
Summary:
Speed up the SAASBO tutorial by using 256 warmup steps and 128 samples (with a comment that we still recommend 512 warmup steps and 256 samples). I also dropped the dimensionality of the problem from 50 to 30 and increased the batch size from 4 to 5. The tutorial now runs in ~12 minutes on Github Actions.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1740

Reviewed By: esantorella

Differential Revision: D44092726

Pulled By: dme65

fbshipit-source-id: 3bd8f627f46394b5f68a1019f4593ad4bade56e9