Actions job to concatenate tutorials data to one CSV and run analysis notebook (#1703)
Summary:
After tutorials performance data is uploaded to artifacts branch, a job runs to concatenate tutorials data to one CSV and run analysis notebook. The notebook, minus cell outputs, and the script that runs it will be in the 'main' branch, while the data and notebook outputs stay in the 'analytics' branch, which continues to not need to stay up-to-date with 'main' since it doesn't have its own Python code.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md#pull-requests)?

Yes

Pull Request resolved: https://github.com/pytorch/botorch/pull/1703

Test Plan:
[x] Ran with manual actions run "on push" (smoke test): [here](https://github.com/pytorch/botorch/actions/runs/4286198759/jobs/7465442757), more recent actions run [here](https://github.com/pytorch/botorch/actions/runs/4388959014/jobs/7686024723).
[x] Auto-generated notebook should have nifty visualizations [here](https://github.com/pytorch/botorch/blob/artifacts/notebooks/tutorials_performance_tracking.ipynb)
[x] Up-to-date dataset of all tutorials runs exists and looks OK [here](https://github.com/pytorch/botorch/blob/artifacts/tutorial_performance_data/all_data.csv)

Reviewed By: saitcakmak

Differential Revision: D43634570

Pulled By: esantorella

fbshipit-source-id: 6d0f2eae16309bc5dbb2c554c35894f28b57c97c