Update google analytics ID
Google Analytics is deprecating the old way of doing things and requires us to set up a new Google Analytics 4 property. Updating the ID in the site config to (hopefully) track site engagement under the new property.
Not sure if this will work with docusaurus v1 but probably easiest to just try (it didn't complain during a local build).