Update google analytics ID (#1752)
Summary:
Google Analytics is deprecating the old way of doing things and requires us to set up a new Google Analytics 4 property. Updating the ID in the site config to (hopefully) track site engagement under the new property. Not sure if this will work with docusaurus v1 but probably easiest to just try (it didn't complain during a local build).

Pull Request resolved: https://github.com/pytorch/botorch/pull/1752

Reviewed By: SebastianAment

Differential Revision: D44299271

Pulled By: Balandat

fbshipit-source-id: 1b84f5c2c936699150856c41edbb410fa3aa188b