Fix typo in saasbo tutorial (#1784)
Summary:
<!--
Thank you for sending the PR! We appreciate you spending the time to make BoTorch better.

Help us understand your motivation by explaining why you decided to make this change.

You can learn more about contributing to BoTorch here: https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md
-->

## Motivation

In the saasbo tutorial it was stated that the posterior is made up of 16 MC samples, which is wrong. There are only 8 samples, since 128 MC steps are performed in combination with a thinning of 8. This amounts to 8 posterior samples.

Furthermore, I noticed that the link to the "custom acquistion tutorial" in the introduction of the notebook is not working, but I do not know where it should link to originally.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/pytorch/botorch/blob/main/CONTRIBUTING.md#pull-requests)?

Yes.

Pull Request resolved: https://github.com/pytorch/botorch/pull/1784

Reviewed By: SebastianAment

Differential Revision: D44773010

Pulled By: Balandat

fbshipit-source-id: 3a2960995f0235ee9fdaeba822e6883fcd3048b7